core: agent path default, legacy via LEGACY_MODE flag
